Rename unit testing namespace
