+++ EngineTestSuite/EngineTestSuite.vcxproj
@@ -198,6 +198,7 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
+    <ClCompile Include="UUIDTests.cpp" />
     <ClCompile Include="WindowManagerTests.cpp" />
   </ItemGroup>
   <ItemGroup>
