+++ EngineTestSuite/EngineTestSuite.vcxproj.filters
@@ -59,5 +59,8 @@
     <ClCompile Include="ConfigurableTests.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="UUIDTests.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
