+++ EngineTestSuite/EngineUtilitiesTests.cpp
@@ -9,20 +9,20 @@
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 
 
-namespace ProjectCoatiTestSuite{
+namespace EngineTestSuite{
 
 	TEST_CLASS(EngineUtilitiesTests) {
 	public:
 		TEST_METHOD(FNV1aEqualityTest) {
-			auto first = EngineUtils::FNV1aHash("ThisIsATest");
-			auto second = EngineUtils::FNV1aHash("ThisIsATest");
+			auto first = EngineCore::FNV1aHash("ThisIsATest");
+			auto second = EngineCore::FNV1aHash("ThisIsATest");
 
 			Assert::AreEqual(first, second);
 		}
 
 		TEST_METHOD(FNV1aInequalityTest) {
-			auto first = EngineUtils::FNV1aHash("ThisIsATest");
-			auto second = EngineUtils::FNV1aHash("SoSoThisIsATest");
+			auto first = EngineCore::FNV1aHash("ThisIsATest");
+			auto second = EngineCore::FNV1aHash("SoSoThisIsATest");
 
 			Assert::AreNotEqual(first, second);
 		}
