+++ EngineTestSuite/UUIDTests.cpp
@@ -8,7 +8,7 @@
 #include "UUID.h"
 
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
+using namespace EngineCore;
 
 namespace EngineTestSuite {
 
@@ -16,7 +16,21 @@ namespace EngineTestSuite {
 		public:
 
 		TEST_METHOD(UUIDConstructorTest) {
+			UUID id("Test");
 
+			Assert::AreNotEqual((unsigned long long)0, id.ID);
+			Assert::AreEqual(std::string("Test"), id.StringID);
+		}
+
+		TEST_METHOD(UUIDComparisonTest) {
+			UUID idA("Foo");
+			UUID idB("Bar");
+			UUID idC("Foo");
+
+			Assert::IsTrue(idA != idB);
+			Assert::IsTrue(idA == idC);
+
+			// Relational operators are available but untested
 		}
 	};
 };
