+++ EngineTestSuite/WindowManagerTests.cpp
@@ -3,7 +3,7 @@
 #include "WindowManager.h"
 using namespace Microsoft::VisualStudio::CppUnitTestFramework;
 
-namespace ProjectCoatiTestSuite
+namespace EngineTestSuite
 {		
 
 	using namespace EngineCore;
